server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
astroid==4.0.4
    # via pylint
autopep8==2.3.2
    # via -r requirements.in
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via x-wr-timezone
dill==0.4.1
    # via pylint
icalendar==7.0.2
    # via
    #   -r requirements.in
    #   recurring-ical-events
    #   x-wr-timezone
idna==3.11
    # via requests
isort==8.0.0
    # via pylint
librt==0.8.1
    # via mypy
mccabe==0.7.0
    # via pylint
mypy==1.19.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
pathspec==1.0.4
    # via mypy
platformdirs==4.9.2
    # via pylint
pycodestyle==2.14.0
    # via autopep8
pylint==4.0.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   icalendar
    #   recurring-ical-events
recurring-ical-events==3.8.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
schedule==1.2.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
tomlkit==0.14.0
    # via pylint
types-icalendar==6.3.2.20260223
    # via -r requirements.in
types-python-dateutil==2.9.0.20260124
    # via
    #   -r requirements.in
    #   types-icalendar
types-pytz==2025.2.0.20251108
    # via types-icalendar
types-requests==2.32.4.20260107
    # via -r requirements.in
typing-extensions==4.15.0
    # via mypy
tzdata==2025.3
    # via
    #   icalendar
    #   recurring-ical-events
    #   x-wr-timezone
urllib3==2.6.3
    # via
    #   requests
    #   types-requests
x-wr-timezone==2.0.1
    # via recurring-ical-events
